Code/ai/requirements.txt
pandas
numpy
scikit-learn
xgboost
fastapi
uvicorn
joblib
requests
flask
python-dotenv
joblib
openai>=1.40.0
